refactor(Signup): Split in files
