+++ packages/client/src/modules/signup/components/Signup/Signup.tsx
@@ -1,0 +1,7 @@
+import Form from "../Form";
+
+function Signup() {
+  return <Form />;
+}
+
+export default Signup;
