+++ packages/client/src/modules/signup/components/Signup/index.tsx
@@ -1,6 +1,1 @@
-import Form from "../Form";
-
-function Signup() {
-  return <Form />;
-}
-export default Signup;
+export { default } from "./Signup";
